some changes to clean code and remove functions from app.vue to helper.ts
