+++ backend/express-session.d.ts
@@ -4,7 +4,7 @@ import { PushSubscription } from 'web-push'
 
 declare module 'express-session' {
   interface SessionData {
-    subscription: PushSubscription
+    subscription?: PushSubscription
     passport?: {
       user?: {
         _id?: Types.ObjectId
